Split MarbleManager from RenderCore. Added render groups for marbles.
