+++ src/client/render/render-core.js
@@ -1,20 +1,17 @@
 import * as THREE from "three";
 import "three/examples/js/loaders/LoaderSupport";
 import "three/examples/js/loaders/GLTFLoader";
-import * as Cookies from "js-cookie";
 import * as Stats from "stats-js";
-import * as config from "../config";
 import { CameraFlyControls } from "./cameras";
 import domReady from "../dom-ready";
 import { levelManager } from "../level-manager";
+import { marbleManager } from "../marble-manager";
 
-let _userData = Cookies.getJSON("user_data");
 const _GLTFLoader = new THREE.GLTFLoader();
 
 let renderCore = function() {
 	let _mainScene = null,
 		_renderer = null,
-		_marbleMeshes = [],
 		_viewport = null, // DOM viewport element
 		_stats = null,
 		_controls = null,
@@ -33,7 +30,7 @@ let renderCore = function() {
 		_stats.begin();
 
 		// Make updates
-		renderCore.updateMarbles();
+		marbleManager.updateMarbles();
 		if(levelManager.activeLevel) levelManager.activeLevel.update();
 
 		// Render the darn thing
@@ -144,107 +141,9 @@ let renderCore = function() {
 
 		getMainScene: function() {
 			return _mainScene;
-		},
-
-		updateMarbles: function() {},
-
-		addMarbleMesh: function(data) {
-			let marbleMesh = new MarbleMesh(data);
-			_marbleMeshes.push(marbleMesh);
-			_mainScene.add(marbleMesh.mesh);
-			_mainScene.add(marbleMesh.nameSprite);
-		},
-
-		removeAllMarbleMeshes: function() {
-			for (let marble of _marbleMeshes) {
-				for (let i = marble.mesh.children.length; i >= 0; i--) {
-					_mainScene.remove(marble.mesh.children[i]);
-				}
-				_mainScene.remove(marble.nameSprite);
-				_mainScene.remove(marble.mesh);
-			}
-
-			_marbleMeshes = [];
-		},
-
-		updateMarbleMeshes: function(newPositions, newRotations, delta) {
-			for (let i = 0; i < _marbleMeshes.length; i++) {
-				// Positions
-				_marbleMeshes[i].mesh.position.x = THREE.Math.lerp(_marbleMeshes[i].mesh.position.x || 0, newPositions[i * 3 + 0], delta);
-				_marbleMeshes[i].mesh.position.y = THREE.Math.lerp(_marbleMeshes[i].mesh.position.y || 0, newPositions[i * 3 + 2], delta);
-				_marbleMeshes[i].mesh.position.z = THREE.Math.lerp(_marbleMeshes[i].mesh.position.z || 0, newPositions[i * 3 + 1], delta);
-
-				// Rotations
-				_marbleMeshes[i].mesh.quaternion.set(
-					newRotations[i * 4 + 0],
-					newRotations[i * 4 + 1],
-					newRotations[i * 4 + 2],
-					newRotations[i * 4 + 3]
-				);
-
-				// Also update the nameSprite position
-				if (_marbleMeshes[i].nameSprite) {
-					_marbleMeshes[i].nameSprite.position.x = (_marbleMeshes[i].mesh.position.x || 0);
-					_marbleMeshes[i].nameSprite.position.y = (_marbleMeshes[i].mesh.position.y || 0) + _marbleMeshes[i].size - .1;
-					_marbleMeshes[i].nameSprite.position.z = (_marbleMeshes[i].mesh.position.z || 0);
-				}
-			}
 		}
 	};
 }();
-
-// Marbles
-const MarbleMesh = function(tags) {
-	this.size = tags.size;
-	this.color = tags.color;
-	this.name = tags.name;
-
-	this.geometry = new THREE.SphereBufferGeometry(this.size, 9, 9);
-	this.materialColor = new THREE.Color(this.color);
-	this.material = new THREE.MeshStandardMaterial({ color: this.materialColor });
-	this.mesh = new THREE.Mesh(this.geometry, this.material);
-
-	// Useful for debugging
-	this.mesh.name = `Marble (${tags.name})`;
-
-	// Shadows
-	this.mesh.castShadow = config.graphics.castShadow.marbles;
-	this.mesh.receiveShadow = config.graphics.receiveShadow.marbles;
-
-	// Highlight own name
-	let nameSpriteOptions = {};
-	if (_userData && _userData.username === this.name) {
-		nameSpriteOptions.color = "#BA0069";
-	}
-
-	// Add name sprite (we avoid parenting, because this will also cause it to inherit the rotation which we do not want)
-	this.nameSprite = makeTextSprite(this.name, nameSpriteOptions);
-};
-
-const makeTextSprite = function(message, options = {}) {
-	let fontFamily = options.fontFamily || "Courier New";
-	let fontSize = 48;
-
-	let canvas = document.createElement("canvas");
-	canvas.width = 512;
-	canvas.height = 128;
-
-	let context = canvas.getContext("2d");
-	context.font = `Bold ${fontSize}px ${fontFamily}`;
-	context.textAlign = "center";
-	context.fillStyle = options.color || "#ffffff";
-	context.fillText(message, 256, fontSize);
-
-	// Canvas contents will be used for a texture
-	let texture = new THREE.Texture(canvas);
-	texture.needsUpdate = true;
-
-	let spriteMaterial = new THREE.SpriteMaterial({ map: texture });
-	let sprite = new THREE.Sprite(spriteMaterial);
-	sprite.scale.set(4, 1, 1.0);
-
-	return sprite;
-};
 
 export {
 	renderCore
